move things and add Readme
