+++ nemcova_data/README.md
@@ -1,4 +1,4 @@
-# CoVital sample data
+# CoVital Nemcova2020 data
 
 This directory contains the data from [Nemcova 2020](https://www.sciencedirect.com/science/article/abs/pii/S1746809420300847), and the data was obtained from [Figshare](https://figshare.com/articles/Monitoring_of_heart_rate_blood_oxygen_saturation_and_blood_pressure_using_smartphone/8198405)
